# Pull Request Description

## Summary
[Provide a brief description of the changes in this PR]

### Issue Reference
Fixes #[Issue Number]

### Motivation and Context
- Why is this change needed?
- What problem does it solve?
- If it fixes an open issue, please link to the issue here

### Dependencies
- List any dependencies that are required for this change
- Include any configuration changes needed
- Note any version updates required

## Type of Change
Please mark the relevant option with an `x`:
- [ ] 🐛 Bug fix (non-breaking change which fixes an issue)
- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 📝 Documentation update (Wiki/README/Code comments)
- [ ] ♻️ Refactor (code improvement without functional changes)
- [ ] 🎨 Style update (formatting, renaming)
- [ ] 🔧 Configuration change
- [ ] 📦 Dependency update

## Testing
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] I have tested this code in the following browsers/environments:
[list environments]

## Quality Checklist
- [ ] I have reviewed my own code before requesting review
- [ ] I have verified there are no other open Pull Requests for the same
update/change
- [ ] All CI/CD pipelines pass without errors or warnings
- [ ] My code follows the established style guidelines of this project
- [ ] I have added necessary documentation (if appropriate)
- [ ] I have commented my code, particularly in complex areas
- [ ] I have made corresponding changes to the README and other relevant
documentation
- [ ] My changes generate no new warnings
- [ ] I have performed a self-review of my own code
- [ ] My code is properly formatted according to project standards

## Screenshots/Recordings (if appropriate)
[Add screenshots or recordings that demonstrate the changes]

## Additional Notes
[Add any additional information that might be helpful for reviewers]

+++ internal/services/datasources/device_and_app_management/utilities/macos_pkg_app_metadata/model.go
@@ -28,7 +28,7 @@ type MetadataDataSourceModel struct {
 	InstallLocation types.String `tfsdk:"install_location"`
 	AppPaths        types.List   `tfsdk:"app_paths"`
 	MinOSVersion    types.String `tfsdk:"min_os_version"`
-	SizeMB          types.Int64  `tfsdk:"size_mb"`
+	SizeMB          types.Int32  `tfsdk:"size_mb"`
 	SHA256Checksum  types.String `tfsdk:"sha256_checksum"`
 	MD5Checksum     types.String `tfsdk:"md5_checksum"`
