+++ internal/services/resources/device_and_app_management/graph_beta/winget_app/model.go
@@ -24,13 +24,13 @@ type WinGetAppResourceModel struct {
 	Owner                         types.String                             `tfsdk:"owner"`
 	Developer                     types.String                             `tfsdk:"developer"`
 	Notes                         types.String                             `tfsdk:"notes"`
-	UploadState                   types.Int64                              `tfsdk:"upload_state"`
+	UploadState                   types.Int32                              `tfsdk:"upload_state"`
 	PublishingState               types.String                             `tfsdk:"publishing_state"`
 	IsAssigned                    types.Bool                               `tfsdk:"is_assigned"`
 	RoleScopeTagIds               types.Set                                `tfsdk:"role_scope_tag_ids"`
-	DependentAppCount             types.Int64                              `tfsdk:"dependent_app_count"`
-	SupersedingAppCount           types.Int64                              `tfsdk:"superseding_app_count"`
-	SupersededAppCount            types.Int64                              `tfsdk:"superseded_app_count"`
+	DependentAppCount             types.Int32                              `tfsdk:"dependent_app_count"`
+	SupersedingAppCount           types.Int32                              `tfsdk:"superseding_app_count"`
+	SupersededAppCount            types.Int32                              `tfsdk:"superseded_app_count"`
 	ManifestHash                  types.String                             `tfsdk:"manifest_hash"`
 	PackageIdentifier             types.String                             `tfsdk:"package_identifier"`
 	AutomaticallyGenerateMetadata types.Bool                               `tfsdk:"automatically_generate_metadata"`
